requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.1.0
appdirs==1.4.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
boto3==1.33.11
botocore==1.33.11
bpemb==0.3.4
browser-cookie3==0.19.1
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
conllu==4.5.3
contourpy==1.2.0
craft==0.2.6
cryptography==41.0.7
curl-cffi==0.5.10
cycler==0.12.1
cymem==2.0.8
datefinder==0.7.3
Deprecated==1.2.14
distro==1.8.0
docker-pycreds==0.4.0
duckling==1.8.0
emoji==2.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
enchant==0.0.1
fake-useragent==1.4.0
filelock==3.13.1
flair==0.13.0
Flask==3.0.0
flask-babel==4.0.0
Flask-Cors==4.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.1
ftfy==6.1.3
gdown==4.7.1
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
html2text==2020.1.16
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-metadata==6.11.0
imutils==0.5.4
inexactsearch==1.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Janome==0.5.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
JPype1==1.4.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
langcodes==3.3.0
langdetect==1.0.9
lark==1.1.8
Levenshtein==0.23.0
lxml==4.9.3
lz4==4.3.2
MailGw-Temporary-Email==0.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
mpld3==0.5.9
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
names==0.3.0
networkx==3.2.1
nltk==3.8.1
np==1.0.2
numpy==1.26.2
openai==1.3.8
opencv-python==4.8.1.78
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
pluggy==1.3.0
pptree==3.1
preshed==3.0.9
protobuf==4.25.1
psutil==5.9.6
pyarrow==14.0.1
pycparser==2.21
pycryptodome==3.19.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pydub==0.25.1
pyenchant==3.2.2
PyExecJS==1.5.1
Pygments==2.17.2
pymailtm==1.1.1
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyparsing==3.1.1
PyPasser==0.0.5
PyPDF2==3.0.1
pypdfium2==4.25.0
pyperclip==1.8.2
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.3
python-dateutil==2.8.2
pytorch_revgrad==0.2.0
pytz==2023.3.post1
PyYAML==6.0.1
random-password-generator==2.2.0
random-username==1.0.2
rapidfuzz==3.5.2
referencing==0.32.0
regex==2023.10.3
reportlab==4.0.7
requests==2.31.0
retrying==1.3.4
rich==13.7.0
rpds-py==0.13.2
s3transfer==0.8.2
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
segtok==1.5.11
selenium==4.16.0
semver==3.0.2
sentencepiece==0.1.99
sentry-sdk==1.38.0
setproctitle==1.3.3
setuptools-scm==8.0.4
silpa-common==0.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sner==0.2.12
sniffio==1.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
spellchecker==0.4
sqlitedict==2.1.0
srsly==2.4.8
stanza==1.7.0
streamlit==1.29.0
sutime==1.0.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==1.1.0
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tls-client==0.2.2
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
transformer-smaller-training-vocab==0.3.3
transformers==4.35.2
trio==0.23.1
trio-websocket==0.11.1
TwoCaptcha==0.0.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
Unidecode==1.3.7
urllib3==1.26.18
validators==0.22.0
wandb==0.16.1
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.12
weasel==0.3.4
websocket-client==1.7.0
Werkzeug==3.0.1
Wikipedia-API==0.6.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
